Rename HTTP.{c,h}pp to HTTPRequest.{c,h}pp
