+++ src/controllers/plugins/PluginController.cpp
@@ -8,7 +8,7 @@
 #    include "controllers/commands/CommandContext.hpp"
 #    include "controllers/commands/CommandController.hpp"
 #    include "controllers/plugins/api/ChannelRef.hpp"
-#    include "controllers/plugins/api/HTTP.hpp"
+#    include "controllers/plugins/api/HTTPRequest.hpp"
 #    include "controllers/plugins/api/IOWrapper.hpp"
 #    include "controllers/plugins/LuaAPI.hpp"
 #    include "controllers/plugins/LuaUtilities.hpp"
